% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_histogram.R
\name{auc_histogram}
\alias{auc_histogram}
\title{Calculate an AUC from a histogram}
\usage{
auc_histogram(x)
}
\arguments{
\item{x}{histogram object from \link[graphics:hist]{graphics::hist}}
}
\description{
Uses a histogram of event probabilties to calculate a precise AUC.  This is
a discrete approximation.  Use this function with many break points with a
large number of data points.
}
\examples{
runif(10000) \%>\%
  hist(breaks = 250) \%>\%
  auc_histogram()
}
\author{
Daniel D. Sjoberg
}
