% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_map.R
\name{count_map}
\alias{count_map}
\title{Checks variable creation for new derived variables at once}
\usage{
count_map(data, checks)
}
\arguments{
\item{data}{data frame}

\item{checks}{list of variables to check.}
}
\description{
Checks variable creation for new derived variables at once
}
\examples{
count_map(
  mtcars,
  list(c("cyl", "am"), c("gear", "carb"))
)
}
