% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfuns-package.R
\docType{package}
\name{hotfun-package}
\alias{hotfun}
\alias{hotfun-package}
\title{hotfun: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
