% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_hotfun}
\alias{print_hotfun}
\alias{print.tbl_ancova}
\title{print and knit_print methods for hotfun objects}
\usage{
\method{print}{tbl_ancova}(x, ...)
}
\arguments{
\item{x}{an object created using hotfun functions}

\item{...}{not used}
}
\description{
print and knit_print methods for hotfun objects
}
\author{
Daniel D. Sjoberg
}
