% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_template}
\alias{project_template}
\title{H.O.T. project template}
\format{
A quoted list defining the H.O.T. project template. Each item of
the list identifies one script or document that appears in the project template.
}
\usage{
project_template
}
\description{
The \code{project_template} object defines the contents of the H.O.T. project
template used in \code{create_hot_project()} and \code{use_hot_file()}.
}
\examples{
\donttest{\dontrun{
create_hot_project(
  path = file.path(tempdir(), "Sjoberg New Project"),
  template = hotfun::project_template
)
}}
}
\seealso{
\code{\link[=create_hot_project]{create_hot_project()}}

\code{\link[=use_hot_file]{use_hot_file()}}
}
\keyword{datasets}
