% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_derived_variables.R
\name{set_derived_variables}
\alias{set_derived_variables}
\title{Apply variable labels to data frame}
\usage{
set_derived_variables(data, path, drop = TRUE)
}
\arguments{
\item{data}{data frame}

\item{path}{file path to the Derived Variables excel file}

\item{drop}{logical indicating whether to drop unlabelled variables}
}
\description{
Takes labels from the Derived Variables excel file and applies them to the
passed data frame.  The function is meant to be used in the pipe.
}
\author{
Daniel D. Sjoberg
}
