% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ancova.R
\name{tbl_ancova}
\alias{tbl_ancova}
\title{Table difference between two groups}
\usage{
tbl_ancova(data, y, x, formula = "{y} ~ {x}", label = NULL,
  method.args = NULL, conf.level = 0.95, hide_n = FALSE,
  estimate_fun = style_sigfig, pvalue_fun = style_pvalue,
  method = "lm")
}
\arguments{
\item{x}{a list of objects of class \code{lm}.  The binary variable must be the first
variable in the covariate list.}
}
\description{
Provided a \link[stats:lm]{stats::lm} object, a table of adjusted differences are returned.
}
