% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotfun-package.R
\docType{package}
\name{hotfun-package}
\alias{hotfun}
\alias{hotfun-package}
\title{hotfun: Collection of Functions Used in the Health Outcomes Team at MSKCC}
\description{
A mixed-bag of utility functions to ease analyses and
    reporting results.
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com}

}
\keyword{internal}
